Move openssl_sha256 module to separate file
